+++ test/base/application_test.dart
@@ -180,14 +180,14 @@ class TestSink extends RequestSink {
 }
 
 class TController extends HTTPController {
-  @httpGet
+  @Bind.get()
   Future<Response> getAll() async {
     return new Response.ok("t_ok");
   }
 }
 
 class RController extends HTTPController {
-  @httpGet
+  @Bind.get()
   Future<Response> getAll() async {
     return new Response.ok("r_ok");
   }