+++ test/base/controller_test.dart
@@ -496,7 +496,7 @@ void main() {
 }
 
 class FilteringController extends HTTPController {
-  @httpGet
+  @Bind.get()
   Future<Response> getAll() async {
     return new Response.ok(null);
   }
@@ -511,28 +511,28 @@ class FilteringController extends HTTPController {
 }
 
 class TController extends HTTPController {
-  @httpGet
+  @Bind.get()
   Future<Response> getAll() async {
     return new Response.ok("getAll");
   }
 
-  @httpGet
-  Future<Response> getOne(@HTTPPath("id") String id) async {
+  @Bind.get()
+  Future<Response> getOne(@Bind.path("id") String id) async {
     return new Response.ok("$id");
   }
 
-  @httpGet
+  @Bind.get()
   Future<Response> getBoth(
-      @HTTPPath("id") String id, @HTTPPath("flag") String flag) async {
+      @Bind.path("id") String id, @Bind.path("flag") String flag) async {
     return new Response.ok("$id$flag");
   }
 
-  @httpPut
-  Future<Response> putOne(@HTTPPath("id") String id) async {
+  @Bind.put()
+  Future<Response> putOne(@Bind.path("id") String id) async {
     throw new Exception("Exception!");
   }
 
-  @httpPost
+  @Bind.post()
   Future<Response> post() async {
     var body = this.request.body.asMap();
 
@@ -541,8 +541,8 @@ class TController extends HTTPController {
 }
 
 class QController extends HTTPController {
-  @httpGet
-  Future<Response> getAll({@HTTPQuery("opt") String opt}) async {
+  @Bind.get()
+  Future<Response> getAll({@Bind.query("opt") String opt}) async {
     if (opt == null) {
       return new Response.ok("NOT");
     }
@@ -550,9 +550,9 @@ class QController extends HTTPController {
     return new Response.ok("OK");
   }
 
-  @httpGet
-  Future<Response> getOne(@HTTPPath("id") String id,
-      {@HTTPQuery("opt") String opt}) async {
+  @Bind.get()
+  Future<Response> getOne(@Bind.path("id") String id,
+      {@Bind.query("opt") String opt}) async {
     if (opt == null) {
       return new Response.ok("NOT");
     }
@@ -562,65 +562,65 @@ class QController extends HTTPController {
 }
 
 class IntController extends HTTPController {
-  @httpGet
-  Future<Response> getOne(@HTTPPath("id") int id) async {
+  @Bind.get()
+  Future<Response> getOne(@Bind.path("id") int id) async {
     return new Response.ok("${id * 2}");
   }
 
-  @httpGet
-  Future<Response> getAll({@HTTPQuery("opt") int opt}) async {
+  @Bind.get()
+  Future<Response> getAll({@Bind.query("opt") int opt}) async {
     return new Response.ok("$opt");
   }
 
-  @httpPost
-  Future<Response> create({@HTTPQuery("opt") int opt}) async {
+  @Bind.post()
+  Future<Response> create({@Bind.query("opt") int opt}) async {
     return new Response.ok("$opt");
   }
 }
 
 class DateTimeController extends HTTPController {
-  @httpGet
-  Future<Response> getOne(@HTTPPath("time") DateTime time) async {
+  @Bind.get()
+  Future<Response> getOne(@Bind.path("time") DateTime time) async {
     return new Response.ok("${time.add(new Duration(seconds: 5))}");
   }
 
-  @httpGet
-  Future<Response> getAll({@HTTPQuery("opt") DateTime opt}) async {
+  @Bind.get()
+  Future<Response> getAll({@Bind.query("opt") DateTime opt}) async {
     return new Response.ok("$opt");
   }
 }
 
 class MultiQueryParamController extends HTTPController {
-  @httpGet
-  Future<Response> get({@HTTPQuery("params") List<String> params}) async {
+  @Bind.get()
+  Future<Response> get({@Bind.query("params") List<String> params}) async {
     return new Response.ok(params.join(","));
   }
 }
 
 class BooleanQueryParamController extends HTTPController {
-  @httpGet
-  Future<Response> get({@HTTPQuery("param") bool param: false}) async {
+  @Bind.get()
+  Future<Response> get({@Bind.query("param") bool param: false}) async {
     return new Response.ok(param ? "true" : "false");
   }
 }
 
 class HTTPParameterController extends HTTPController {
   @requiredHTTPParameter
-  @HTTPHeader("X-Request-id")
+  @Bind.header("X-Request-id")
   String requestId;
   @requiredHTTPParameter
-  @HTTPQuery("Shaqs")
+  @Bind.query("Shaqs")
   int numberOfShaqs;
-  @HTTPHeader("Location")
+  @Bind.header("Location")
   String location;
-  @HTTPQuery("number")
+  @Bind.query("number")
   int number;
 
-  @httpGet
-  Future<Response> get(@HTTPHeader("Cookie") String cookieBrand,
-      @HTTPQuery("Table") String tableBrand,
-      {@HTTPHeader("Milk") String milkBrand,
-      @HTTPQuery("table_legs") int numberOfTableLegs}) async {
+  @Bind.get()
+  Future<Response> get(@Bind.header("Cookie") String cookieBrand,
+      @Bind.query("Table") String tableBrand,
+      {@Bind.header("Milk") String milkBrand,
+      @Bind.query("table_legs") int numberOfTableLegs}) async {
     return new Response.ok({
       "location": location,
       "x-request-id": requestId,
@@ -635,8 +635,8 @@ class HTTPParameterController extends HTTPController {
 }
 
 class ModelEncodeController extends HTTPController {
-  @httpGet
-  Future<Response> getThings(@HTTPPath("thing") String thing) async {
+  @Bind.get()
+  Future<Response> getThings(@Bind.path("thing") String thing) async {
     if (thing == "list") {
       return new Response.ok([
         {"id": 1},
@@ -671,8 +671,8 @@ class ModelEncodeController extends HTTPController {
 }
 
 class ContentTypeController extends HTTPController {
-  @httpGet
-  Future<Response> getThing(@HTTPQuery("opt") String opt) async {
+  @Bind.get()
+  Future<Response> getThing(@Bind.query("opt") String opt) async {
     if (opt == "responseContentType") {
       responseContentType = new ContentType("text", "plain");
       return new Response.ok("body");
@@ -686,9 +686,9 @@ class ContentTypeController extends HTTPController {
 }
 
 class DuplicateParamController extends HTTPController {
-  @httpGet
-  Future<Response> getThing(@HTTPQuery("list") List<String> list,
-      @HTTPQuery("single") String single) async {
+  @Bind.get()
+  Future<Response> getThing(@Bind.query("list") List<String> list,
+      @Bind.query("single") String single) async {
     return new Response.ok({"list": list, "single": single});
   }
 }
@@ -696,12 +696,12 @@ class DuplicateParamController extends HTTPController {
 class DecodeCallbackController extends HTTPController {
   bool didDecode = false;
 
-  @httpGet
+  @Bind.get()
   Future<Response> getThing() async {
     return new Response.ok({"didDecode": didDecode});
   }
 
-  @httpPost
+  @Bind.post()
   Future<Response> postThing() async {
     return new Response.ok({"didDecode": didDecode});
   }