+++ test/base/isolate/recovery_test.dart
@@ -81,7 +81,7 @@ class TestSink extends RequestSink {
 }
 
 class UncaughtCrashController extends HTTPController {
-  @httpGet
+  @Bind.get()
   Future<Response> crashUncaught() async {
     new Future(() {
       var x;
@@ -90,8 +90,8 @@ class UncaughtCrashController extends HTTPController {
     return new Response.ok(null);
   }
 
-  @httpGet
-  Future<Response> dontCrash(@HTTPPath("id") int id) async {
+  @Bind.get()
+  Future<Response> dontCrash(@Bind.path("id") int id) async {
     return new Response.ok(null);
   }
 }